+++ sdk/servicebus/azure_messaging_servicebus/src/common/authorizer.rs
@@ -16,7 +16,6 @@ use azure_core_amqp::{
 use rand::{rng, RngExt};
 use std::{
     collections::HashMap,
-    pin::Pin,
     sync::{Arc, Mutex as SyncMutex, OnceLock, Weak},
 };
 use tracing::{debug, trace, warn};
@@ -47,7 +46,7 @@ impl Default for TokenRefreshTimes {
 
 pub(crate) struct Authorizer {
     authorization_scopes: AsyncMutex<HashMap<Url, AccessToken>>,
-    authorization_refresher: OnceLock<Pin<Box<dyn SpawnedTask>>>,
+    authorization_refresher: OnceLock<SpawnedTask>,
     /// Bias to apply to token refresh time. This determines how much time we will refresh the token before it expires.
     token_refresh_bias: SyncMutex<TokenRefreshTimes>,
     credential: Arc<dyn TokenCredential>,